Ajustando a captura e detecção
ajustando algumas logicas de detecção e alguns erro em relação a conexão.
